+++ editors/vscode/test/grammar/special-ops.lisp
@@ -37,6 +37,9 @@
 (ignore-errors (risky-operation))
 ;^^^^^^^^^^^^^ keyword.control.elps
 
+(unwind-protect (risky-operation) (cleanup))
+;^^^^^^^^^^^^^^ keyword.control.elps
+
 (dotimes (i 10) (debug-print i))
 ;^^^^^^^ keyword.control.elps